+++ iviz_msgs/std_msgs/msg/Float64.cs
@@ -33,7 +33,7 @@ namespace Iviz.Msgs.StdMsgs
         
         Float64 IDeserializable<Float64>.RosDeserialize(ref Buffer b)
         {
-            return new Float64(ref b);
+            return new(ref b);
         }
     
         public void RosSerialize(ref Buffer b)
