+++ iviz_msgs/std_msgs/msg/Int64.cs
@@ -33,7 +33,7 @@ namespace Iviz.Msgs.StdMsgs
         
         Int64 IDeserializable<Int64>.RosDeserialize(ref Buffer b)
         {
-            return new Int64(ref b);
+            return new(ref b);
         }
     
         public void RosSerialize(ref Buffer b)
