+++ iviz_msgs/std_msgs/msg/Int8.cs
@@ -33,7 +33,7 @@ namespace Iviz.Msgs.StdMsgs
         
         Int8 IDeserializable<Int8>.RosDeserialize(ref Buffer b)
         {
-            return new Int8(ref b);
+            return new(ref b);
         }
     
         public void RosSerialize(ref Buffer b)
