+++ iviz_msgs/std_msgs/msg/UInt16.cs
@@ -33,7 +33,7 @@ namespace Iviz.Msgs.StdMsgs
         
         UInt16 IDeserializable<UInt16>.RosDeserialize(ref Buffer b)
         {
-            return new UInt16(ref b);
+            return new(ref b);
         }
     
         public void RosSerialize(ref Buffer b)
