+++ iviz_msgs/std_msgs/msg/UInt32.cs
@@ -33,7 +33,7 @@ namespace Iviz.Msgs.StdMsgs
         
         UInt32 IDeserializable<UInt32>.RosDeserialize(ref Buffer b)
         {
-            return new UInt32(ref b);
+            return new(ref b);
         }
     
         public void RosSerialize(ref Buffer b)
