+++ iviz_msgs/std_msgs/msg/UInt64.cs
@@ -33,7 +33,7 @@ namespace Iviz.Msgs.StdMsgs
         
         UInt64 IDeserializable<UInt64>.RosDeserialize(ref Buffer b)
         {
-            return new UInt64(ref b);
+            return new(ref b);
         }
     
         public void RosSerialize(ref Buffer b)
